feat: create event preview cards & page for dynamic routes
The API to provide event data is not ready at the moment, so I'm using some dummy data in the file temporarily
to give the cards something to display.

Right now the display uses flexbox, but we might consider using grid depending on how many events we want
to display at once.