Date offset fix (#192)
* fix: correct time offset

Due to upstream issues parsing Japanese event dates and times with UTC,
all events are offset +9 hours. Created a helper function (using moment)
to set times and dates back the appropriate amount.

* fix: correct time offset in table view in timeline

* fix: correct time offset in groups of dates

In order to correct the time offset problem (see previous commit comments),
in the grouping of events, it's necessary to offset the time by 9 hours
in the getUniqueDates as well.

* fix: update page title and Hero to match current name

* fix: temp hide inaccurate date, time ranges

* fix: delete hyphens from navbar title

* fix: delete hyphens from navbar title

* fix: change client route ref name from "/events" to "/event"

* fix: use getTimeRange func